Expand probes if they have mixed args
Instead of erroring if a probe utilizes
the `args` or `retval` builtins and they
are of multiple types depending on the provider
or attachpoint, separate them into separate
ASTs/programs earlier in the pass pipeline.

This also cleans up a lot of complicated code
for classic tracepoints by removing the is_tparg
state variable and all the logic for splitting
out mixed arg tracepoints in codegen.

stack-info: PR: https://github.com/bpftrace/bpftrace/pull/4660, branch: user/jordalgo/tp_args_v2/1
