+++ src/ast/ast.cpp
@@ -374,7 +374,7 @@ void AttachPoint::set_index(int index)
 
 std::string Probe::args_typename() const
 {
-  return "struct " + orig_name + "_args";
+  return "struct " + orig_name + attach_points.front()->func + "_args";
 }
 
 int Probe::index() const
