+++ src/ast/passes/field_analyser.cpp
@@ -2,12 +2,13 @@
 #include <cassert>
 
 #include "arch/arch.h"
+#include "ast/passes/ap_expansion.h"
 #include "ast/passes/field_analyser.h"
-#include "ast/passes/probe_expansion.h"
 #include "ast/visitor.h"
 #include "bpftrace.h"
 #include "dwarf_parser.h"
 #include "probe_matcher.h"
+#include "tracepoint_format_parser.h"
 #include "util/strings.h"
 
 namespace bpftrace::ast {
@@ -89,13 +90,11 @@ void FieldAnalyser::visit(Builtin &builtin)
   } else if (builtin.ident == "args") {
     if (!probe_)
       return;
-    resolve_args(*probe_);
     has_builtin_args_ = true;
     return;
   } else if (builtin.ident == "__builtin_retval") {
     if (!probe_)
       return;
-    resolve_args(*probe_);
 
     const auto *arg = bpftrace_.structs.GetProbeArg(*probe_, RETVAL_FIELD_NAME);
     if (arg)
@@ -219,110 +218,6 @@ void FieldAnalyser::visit(Unop &unop)
     auto tmp = *sized_type_.GetPointeeTy();
     sized_type_ = std::move(tmp);
     resolve_fields(sized_type_);
-  }
-}
-
-void FieldAnalyser::resolve_args(Probe &probe)
-{
-  for (auto *ap : probe.attach_points) {
-    // load probe arguments into a special record type "struct <probename>_args"
-    std::shared_ptr<Struct> probe_args;
-
-    auto probe_type = probetype(ap->provider);
-    if (probe_type != ProbeType::fentry && probe_type != ProbeType::fexit &&
-        probe_type != ProbeType::rawtracepoint &&
-        probe_type != ProbeType::uprobe)
-      continue;
-
-    if (expansions_.get_expansion(*ap) != ExpansionType::NONE) {
-      std::set<std::string> matches;
-
-      // Find all the matches for the wildcard..
-      try {
-        matches = bpftrace_.probe_matcher_->get_matches_for_ap(*ap);
-      } catch (const WildcardException &e) {
-        probe.addError() << e.what();
-        return;
-      }
-
-      // ... and check if they share same arguments.
-
-      std::shared_ptr<Struct> ap_args;
-      for (const auto &match : matches) {
-        // Both uprobes and fentry have a target (binary for uprobes, kernel
-        // module for fentry).
-        std::string func = match;
-        std::string target = util::erase_prefix(func);
-        std::string err;
-
-        // Trying to attach to multiple fentry. If some of them fails on
-        // argument resolution, do not fail hard, just print a warning and
-        // continue with other functions.
-        if (probe_type == ProbeType::fentry || probe_type == ProbeType::fexit) {
-          ap_args = bpftrace_.btf_->resolve_args(
-              func, probe_type == ProbeType::fexit, true, false, err);
-
-        } else if (probe_type == ProbeType::rawtracepoint) {
-          ap_args = bpftrace_.btf_->resolve_raw_tracepoint_args(func, err);
-        } else { // uprobe
-          Dwarf *dwarf = bpftrace_.get_dwarf(target);
-          if (dwarf)
-            ap_args = dwarf->resolve_args(func);
-          else
-            ap->addWarning() << "No debuginfo found for " << target;
-        }
-
-        if (!ap_args) {
-          ap->addWarning() << probetypeName(probe_type) << ap->func << ": "
-                           << err;
-          continue;
-        }
-
-        if (!probe_args)
-          probe_args = ap_args;
-        else if (*ap_args != *probe_args) {
-          ap->addError() << "Probe has attach points with mixed arguments";
-          break;
-        }
-      }
-    } else {
-      std::string err;
-      // Resolving args for an explicit function failed, print an error and fail
-      if (probe_type == ProbeType::fentry || probe_type == ProbeType::fexit) {
-        probe_args = bpftrace_.btf_->resolve_args(
-            ap->func, probe_type == ProbeType::fexit, true, false, err);
-
-      } else if (probe_type == ProbeType::rawtracepoint) {
-        probe_args = bpftrace_.btf_->resolve_raw_tracepoint_args(ap->func, err);
-      } else { // uprobe
-        Dwarf *dwarf = bpftrace_.get_dwarf(ap->target);
-        if (dwarf) {
-          probe_args = dwarf->resolve_args(ap->func);
-        } else {
-          ap->addWarning() << "No debuginfo found for " << ap->target;
-        }
-        if (probe_args &&
-            probe_args->fields.size() >= arch::Host::arguments().size()) {
-          ap->addError() << "\'args\' builtin is not supported for "
-                         << "probes with stack-passed arguments.";
-        }
-      }
-
-      if (!probe_args) {
-        ap->addError() << probetypeName(probe_type) << ap->func << ": " << err;
-        return;
-      }
-    }
-
-    // check if we already stored arguments for this probe
-    auto args = bpftrace_.structs.Lookup(probe.args_typename()).lock();
-    if (args && *args != *probe_args) {
-      // we did, and it's different...trigger the error
-      ap->addError() << "Probe has attach points with mixed arguments";
-    } else {
-      // store/save args for each ap for later processing
-      bpftrace_.structs.Add(probe.args_typename(), std::move(probe_args));
-    }
   }
 }
