+++ tests/CMakeLists.txt
@@ -55,7 +55,7 @@ add_executable(bpftrace_test
   output.cpp
   parser.cpp
   portability_analyser.cpp
-  probe_expansion.cpp
+  ap_expansion.cpp
   procmon.cpp
   probe.cpp
   config_analyser.cpp
