--- tests/codegen/args_multiple_tracepoints_category_wild.cpp
@@ -1,14 +1,0 @@
-#include "../mocks.h"
-#include "common.h"
-
-using ::testing::_;
-using ::testing::Return;
-
-namespace bpftrace::test::codegen {
-
-TEST(codegen, args_multiple_tracepoints_category_wild)
-{
-  test("tracepoint:sched*:sched_* { @[args.common_field] = 1; }", NAME);
-}
-
-} // namespace bpftrace::test::codegen
