+++ tests/codegen/args_multiple_tracepoints_wild.cpp
@@ -8,7 +8,7 @@ namespace bpftrace::test::codegen {
 
 TEST(codegen, args_multiple_tracepoints_wild)
 {
-  test("tracepoint:sched:sched_* { @[args.common_field] = 1; }", NAME);
+  test("tracepoint:sched:sched_on* { @[args.common_field] = 1; }", NAME);
 }
 
 } // namespace bpftrace::test::codegen
