+++ tests/codegen/llvm/args_multiple_tracepoints.ll
@@ -43,12 +43,12 @@ declare void @llvm.lifetime.start.p0(i64 immarg %0, ptr nocapture %1) #2
 declare void @llvm.lifetime.end.p0(i64 immarg %0, ptr nocapture %1) #2
 
 ; Function Attrs: nounwind
-define i64 @tracepoint_sched_sched_two_2(ptr %0) #0 section "s_tracepoint_sched_sched_two_2" !dbg !56 {
+define i64 @tracepoint_sched_sched_one_twin_2(ptr %0) #0 section "s_tracepoint_sched_sched_one_twin_2" !dbg !56 {
 entry:
   %"@_val" = alloca i64, align 8
   %"@_key" = alloca i64, align 8
   %1 = call ptr @llvm.preserve.static.offset(ptr %0)
-  %2 = getelementptr i8, ptr %1, i64 16
+  %2 = getelementptr i8, ptr %1, i64 8
   %3 = load volatile i64, ptr %2, align 8
   call void @llvm.lifetime.start.p0(i64 -1, ptr %"@_key")
   store i64 %3, ptr %"@_key", align 8
@@ -123,6 +123,6 @@ attributes #2 = { nocallback nofree nosync nounwind willreturn memory(argmem: re
 !53 = !DIDerivedType(tag: DW_TAG_pointer_type, baseType: !4, size: 64)
 !54 = !{!55}
 !55 = !DILocalVariable(name: "ctx", arg: 1, scope: !50, file: !2, type: !53)
-!56 = distinct !DISubprogram(name: "tracepoint_sched_sched_two_2", linkageName: "tracepoint_sched_sched_two_2", scope: !2, file: !2, type: !51, flags: DIFlagPrototyped, spFlags: DISPFlagDefinition, unit: !46, retainedNodes: !57)
+!56 = distinct !DISubprogram(name: "tracepoint_sched_sched_one_twin_2", linkageName: "tracepoint_sched_sched_one_twin_2", scope: !2, file: !2, type: !51, flags: DIFlagPrototyped, spFlags: DISPFlagDefinition, unit: !46, retainedNodes: !57)
 !57 = !{!58}
 !58 = !DILocalVariable(name: "ctx", arg: 1, scope: !56, file: !2, type: !53)
