+++ tests/codegen/llvm/args_multiple_tracepoints_wild.ll
@@ -42,24 +42,6 @@ declare void @llvm.lifetime.start.p0(i64 immarg %0, ptr nocapture %1) #2
 ; Function Attrs: nocallback nofree nosync nounwind willreturn memory(argmem: readwrite)
 declare void @llvm.lifetime.end.p0(i64 immarg %0, ptr nocapture %1) #2
 
-; Function Attrs: nounwind
-define i64 @tracepoint_sched_sched_two_2(ptr %0) #0 section "s_tracepoint_sched_sched_two_2" !dbg !56 {
-entry:
-  %"@_val" = alloca i64, align 8
-  %"@_key" = alloca i64, align 8
-  %1 = call ptr @llvm.preserve.static.offset(ptr %0)
-  %2 = getelementptr i8, ptr %1, i64 16
-  %3 = load volatile i64, ptr %2, align 8
-  call void @llvm.lifetime.start.p0(i64 -1, ptr %"@_key")
-  store i64 %3, ptr %"@_key", align 8
-  call void @llvm.lifetime.start.p0(i64 -1, ptr %"@_val")
-  store i64 1, ptr %"@_val", align 8
-  %update_elem = call i64 inttoptr (i64 2 to ptr)(ptr @AT_, ptr %"@_key", ptr %"@_val", i64 0)
-  call void @llvm.lifetime.end.p0(i64 -1, ptr %"@_val")
-  call void @llvm.lifetime.end.p0(i64 -1, ptr %"@_key")
-  ret i64 1
-}
-
 attributes #0 = { nounwind }
 attributes #1 = { nocallback nofree nosync nounwind speculatable willreturn memory(none) }
 attributes #2 = { nocallback nofree nosync nounwind willreturn memory(argmem: readwrite) }
@@ -123,6 +105,3 @@ attributes #2 = { nocallback nofree nosync nounwind willreturn memory(argmem: re
 !53 = !DIDerivedType(tag: DW_TAG_pointer_type, baseType: !4, size: 64)
 !54 = !{!55}
 !55 = !DILocalVariable(name: "ctx", arg: 1, scope: !50, file: !2, type: !53)
-!56 = distinct !DISubprogram(name: "tracepoint_sched_sched_two_2", linkageName: "tracepoint_sched_sched_two_2", scope: !2, file: !2, type: !51, flags: DIFlagPrototyped, spFlags: DISPFlagDefinition, unit: !46, retainedNodes: !57)
-!57 = !{!58}
-!58 = !DILocalVariable(name: "ctx", arg: 1, scope: !56, file: !2, type: !53)
