+++ tests/codegen/regression.cpp
@@ -26,12 +26,14 @@ TEST(codegen, regression_957)
                 .add(ast::CreateCheckAttachpointsPass())
                 .add(ast::CreateControlFlowPass())
                 .add(ast::CreateMacroExpansionPass())
+                .add(ast::CreateApExpansionPass())
                 .add(ast::CreateProbeExpansionPass())
                 .add(ast::CreateFieldAnalyserPass())
                 .add(ast::CreateNamedParamsPass())
                 .add(ast::CreateMapSugarPass())
                 .add(ast::CreateLLVMInitPass())
                 .add(ast::CreateClangBuildPass())
+                .add(ast::CreateProbeExpansionPass({ ProbeType::tracepoint }))
                 .add(ast::CreateTypeSystemPass())
                 .add(ast::CreateSemanticPass())
                 .add(ast::CreateResourcePass())
