+++ tests/parser.cpp
@@ -1,11 +1,11 @@
 #include <climits>
 #include <sstream>
 
+#include "ast/passes/ap_expansion.h"
 #include "ast/passes/attachpoint_passes.h"
 #include "ast/passes/c_macro_expansion.h"
 #include "ast/passes/clang_parser.h"
 #include "ast/passes/printer.h"
-#include "ast/passes/probe_expansion.h"
 #include "driver.h"
 #include "gtest/gtest.h"
