+++ tests/probe.cpp
@@ -1,3 +1,4 @@
+#include "ast/passes/ap_expansion.h"
 #include "ast/passes/attachpoint_passes.h"
 #include "ast/passes/clang_build.h"
 #include "ast/passes/codegen_llvm.h"
@@ -40,10 +41,11 @@ void gen_bytecode(const std::string &input, std::stringstream &out)
                 .add(ast::CreateResolveImportsPass())
                 .add(ast::CreateParseAttachpointsPass())
                 .add(ast::CreateControlFlowPass())
-                .add(ast::CreateProbeExpansionPass())
+                .add(ast::CreateApExpansionPass())
                 .add(ast::CreateMacroExpansionPass())
                 .add(CreateParseBTFPass())
                 .add(ast::CreateMapSugarPass())
+                .add(ast::CreateProbeExpansionPass())
                 .add(ast::CreateFieldAnalyserPass())
                 .add(ast::CreateNamedParamsPass())
                 .add(ast::CreateLLVMInitPass())
